anycompress: a layer for transparently compressing and caching typeURLs
A tm-db compatible .DB that when given the constructor arguments for a database, acts as a pass through but intercepts
protobuf serialized types.Any's and extracts the typeURL, hashes it and then serializes the value. On .Get, we examine retrieved values from the underlying database and transparently re-convert to the proto of a *types.Any if needed.

To perform this magic, we append 6 bytes to every *types.Any value consisting of:
value_in_types.Any.Value + 32bit FNV hash of typeURL + magic suffix of "\xfe\xff".
The purpose of the magic suffix is to help with identification later on of values that we should intercept.

Fixes #6193